requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.11.0
bleach==6.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==3.2.22
django-appconf==1.0.5
django-compressor==4.4
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.3
django-mail-templated==2.6.5
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-robots==6.1
django-simple-captcha==0.5.20
django-summernote==0.8.20.0
django-taggit==4.0.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
drf-yasg==1.21.7
flake8==6.1.0
idna==3.6
importlib-metadata==6.8.0
inflection==0.5.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
Markdown==3.5
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
Pillow==10.0.1
platformdirs==4.1.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
rcssmin==1.1.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.1
rpds-py==0.10.6
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
swagger-spec-validator==3.0.3
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
webencodings==0.5.1
zipp==3.17.0
